made simple libc.a so we don't repeat code everywhere :p
