+++ src/libc/src/crt.asm
@@ -1,11 +1,4 @@
-global intsyscall
-
-intsyscall:
-  int 0x45
-  ret
-
 global _start
-
 extern main
 
 _start:
