+++ src/libc/src/include/sys/mandelbrot.h
@@ -1,0 +1,39 @@
+#ifndef __SYS_MANDELBROT_H__
+#define __SYS_MANDELBROT_H__
+
+#include <stddef.h>
+#include <stdint.h>
+
+#define SYSCALL_OPEN 0
+#define SYSCALL_CLOSE 1
+#define SYSCALL_READ 2
+#define SYSCALL_WRITE 3
+#define SYSCALL_EXEC 4
+#define SYSCALL_MMAP 5
+#define SYSCALL_MUNMAP 6
+#define SYSCALL_STAT 7
+#define SYSCALL_FSTAT 8
+#define SYSCALL_GETPID 9
+#define SYSCALL_EXIT 10
+#define SYSCALL_FORK 11
+#define SYSCALL_GETTIMEOFDAY 12
+#define SYSCALL_FSYNC 13
+#define SYSCALL_IOCTL 14
+
+#define IOCTL_FBDEV_GET_WIDTH 1
+#define IOCTL_FBDEV_GET_HEIGHT 2
+#define IOCTL_FBDEV_GET_BPP 3
+
+#define PROT_READ 0x1
+#define PROT_WRITE 0x2
+#define PROT_EXEC 0x4
+#define PROT_NONE 0x0
+
+#define MAP_PRIVATE 0x1
+#define MAP_SHARED 0x2
+#define MAP_FIXED 0x4
+#define MAP_ANON 0x8
+
+uint64_t intsyscall(uint64_t id, uint64_t arg_1, uint64_t arg_2, uint64_t arg_3, uint64_t arg_4, uint64_t arg_5);
+
+#endif
