+++ src/libc/src/sys/mandelbrot.asm
@@ -1,0 +1,5 @@
+global intsyscall
+
+intsyscall:
+  int 0x45
+  ret
