+++ src/prog/mandelbrot/Makefile
@@ -6,6 +6,7 @@ AS = nasm
 
 CFLAGS := $(CFLAGS) \
 	-Isrc/include \
+	-I../../libc/src/include \
 	-ffreestanding \
 	-mno-red-zone \
 	-fno-pic -no-pie \
@@ -16,6 +17,7 @@ ASFLAGS := $(ASFLAGS) \
 
 LDFLAGS := \
 	-Tlinker.ld \
+	-L../../../build/libc \
 
 CFILES := $(shell find src/ -name '*.c')
 ASFILES := $(shell find src/ -name '*.asm')
@@ -30,7 +32,7 @@ compile: ld
 	
 ld: $(OFILES)
 	@ echo "[LD] $^"
-	@ $(LD) $(LDFLAGS) $^ -o $(TARGET)
+	@ $(LD) $(LDFLAGS) $^ -lc -o $(TARGET)
 
 %.o: %.c
 	@ echo "[CC] $<"
