--- src/prog/sin/src/main.asm
@@ -1,14 +1,0 @@
-global intsyscall
-
-intsyscall:
-  int 0x45
-  ; jmp $
-  ret
-
-global _start
-
-extern main
-
-_start:
-  call main
-  jmp $
