+++ src/prog/sin/src/sin.c
