+++ src/prog/trig/src/trig.c
